refactor: Modalのhooksのconditionallyのアラートに対応
