+++ src/app/container/tablet/Modal/Modal/Modal.tsx
@@ -27,7 +27,7 @@ import { SelectTags } from '../Modals/SelectTags/SelectTags';
 import { ManageTags } from '../../Drawer/ManageTags/ManageTags';
 import { SettingUserInfo } from '../../Drawer/ManageUserInfo/ManageUserInfo';
 import { DeleteAccountForm } from '../../Drawer/DeleteAccountForm/DeleteAccountForm';
-import { useModalSize, medium } from './context/useModalSize';
+import { useModalStyle, medium } from './context/useModalStyle';
 import { StyledDialog } from './components/StyledDialog';
 import { ManageInstagramAccounts } from '../../Drawer/ManageInstagramAccounts/ManageInstagmaAccounts';
 import { SelectInstagramAccounts } from '../Modals/SelectInstagramAccounts/SelectInstagramAccounts';
@@ -72,7 +72,7 @@ export const useModalProps = () => {
 
 type Props = ReturnType<typeof useModalProps>;
 
-const useStyles = makeStyles((theme) => {
+const useStyles = makeStyles(() => {
   return createStyles({
     root: {
       padding: 0,
@@ -90,26 +90,28 @@ export const ModalPresenter: React.FC<Props> = (props) => {
   let modalContent: React.ReactNode = () => <></>;
 
   // modalStyleにモーダルの表示形式の設定。サイズやoverflowなどのプロパティを設定する。デフォルトはlarge
-  let modalStyle = useModalSize('large');
-  const modalStyleMobile =
-    props.setModal === 'google_search'
-      ? useModalSize('upperSide')
-      : useModalSize('fullScreen');
+  let modalStyle = useModalStyle('large');
+
+  const fullScreenModalStyle = useModalStyle('fullScreen');
+  const upperSideModalStyle = useModalStyle('upperSide');
+  const propsModalStyle = useModalStyle(props.modalSize);
+  const mediumModalStyle = useModalStyle('medium');
+  const currentModalStyle = useModalStyle(props.currentModalContent.modalSize);
 
   switch (props.setModal) {
     case 'content_modal':
       modalContent = <ContentModal />;
       break;
     case 'footer_item_modal':
-      modalStyle = useModalSize(props.currentModalContent.modalSize);
+      modalStyle = currentModalStyle;
       modalContent = <FooterItemModal />;
       break;
     case 'google_search':
-      modalStyle = useModalSize('upperSide');
+      modalStyle = upperSideModalStyle;
       modalContent = <GoogleSearch />;
       break;
     case 'instagram_media_modal':
-      modalStyle = useModalSize(props.currentModalContent.modalSize);
+      modalStyle = currentModalStyle;
       modalContent = <InstagramMediaModal />;
       break;
     case 'select_tags':
@@ -127,18 +129,18 @@ export const ModalPresenter: React.FC<Props> = (props) => {
       modalContent = <ArticleEditor />;
       break;
     case 'edit_footer_item':
-      modalStyle = useModalSize(props.modalSize);
+      modalStyle = propsModalStyle;
       modalContent = <FooterItemEditor />;
       break;
     case 'edit_tags':
-      modalStyle = useModalSize('medium');
+      modalStyle = mediumModalStyle;
       modalContent = <ManageTags />;
       break;
     case 'manage_instagram':
       modalContent = <ManageInstagramAccounts />;
       break;
     case 'setting_theme':
-      modalStyle = useModalSize('medium');
+      modalStyle = mediumModalStyle;
       modalContent = <ManageTheme />;
       break;
     case 'setting_user_info':
@@ -151,7 +153,7 @@ export const ModalPresenter: React.FC<Props> = (props) => {
       modalContent = <DeleteAccountForm />;
       break;
     case 'popup_not_email_verified':
-      modalStyle = useModalSize('medium');
+      modalStyle = mediumModalStyle;
       modalContent = <PageNotEmailVerified />;
       break;
 
@@ -184,7 +186,11 @@ export const ModalPresenter: React.FC<Props> = (props) => {
       setModal={props.setModal}
       isEditting={props.edittingPrams.isEditting}
       modalStyle={modalStyle}
-      modalStyleMobile={modalStyleMobile}
+      modalStyleMobile={
+        props.setModal === 'google_search'
+          ? upperSideModalStyle
+          : fullScreenModalStyle
+      }
       className={classes.root}
       open={props.isModalOpen}
       TransitionComponent={Transition}
@@ -210,7 +216,7 @@ export const ModalPresenter: React.FC<Props> = (props) => {
   );
 };
 
-export const Modal = () => {
+export const Modal = (): JSX.Element => {
   const props = useModalProps();
 
   return <ModalPresenter {...props} />;
